requirements.txt
discord.py>=2.0.0
yt-dlp>=2025.7.6